demos/GUIAgent/ShowUIServices/requirements.txt
fastapi
uvicorn
transformers>=4.45.0
torch
torchvision
accelerate
pillow
qwen-vl-utils
python-multipart
numpy
